backend/requirements.txt
fastapi
uvicorn
streamlit
pinecone-client
sentence-transformers
python-dotenv
